# openapi-oficina.yaml
openapi: 3.0.3
info:
  title: API Oficina Mecânica
  version: "v1"
  description: |
    API em Python (Flask) para gestão de autenticação de usuários (JWT), busca externa
    de produtos (token de serviço) e carrinho por usuário. Documentação em PT-BR.
  contact:
    name: Suporte da API
    email: suporte@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Desenvolvimento local
...
security:
  - bearerAuth: []
paths:
...
  /produto_detalhes:
    get:
      tags: [Produto]
      summary: Detalhes de produto e similares
      parameters:
        - in: query
          name: id
          schema: { type: integer }
        - in: query
          name: nomeProduto
          schema: { type: string }
        - in: query
          name: codigoReferencia
          schema: { type: string }
        - in: query
          name: marca
          schema: { type: string }
      responses:
        '200': { description: Item principal e similares }
        '400': { $ref: '#/components/responses/Resp400' }
  /carrinho:
    get:
      tags: [Carrinho]
      summary: Listar itens do carrinho
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de itens do carrinho }
        '401': { $ref: '#/components/responses/Resp401' }
  /salvar_produto:
    post:
      tags: [Carrinho]
      summary: Adicionar produto ao carrinho
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_api_externa]
              properties:
                id_api_externa: { type: integer }
                nome: { type: string }
                codigo_referencia: { type: string }
                url_imagem: { type: string }
                preco_original: { type: number }
                preco_final: { type: number }
                desconto: { type: number }
                marca: { type: string }
                quantidade: { type: integer, default: 1 }
      responses:
        '200': { description: Atualizado (incrementou quantidade) }
        '201': { description: Criado }
        '400': { $ref: '#/components/responses/Resp400' }
        '401': { $ref: '#/components/responses/Resp401' }
        '500': { $ref: '#/components/responses/Resp500' }
  /carrinho/produto/remover:
    post:
      tags: [Carrinho]
      summary: Remover produto do carrinho
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_api_externa]
              properties:
                id_api_externa: { type: integer }
      responses:
        '200': { description: Removido }
        '400': { $ref: '#/components/responses/Resp400' }
        '401': { $ref: '#/components/responses/Resp401' }
  /carrinho/produto/atualizar-quantidade:
    post:
      tags: [Carrinho]
      summary: Atualizar quantidade de um produto no carrinho
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_api_externa, quantidade]
              properties:
                id_api_externa: { type: integer }
                quantidade: { type: integer, minimum: 0 }
      responses:
        '200': { description: Atualizado }
        '400': { $ref: '#/components/responses/Resp400' }
        '401': { $ref: '#/components/responses/Resp401' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Resp400: { description: Requisição inválida }
    Resp401: { description: Não autorizado }
...
    Resp500: { description: Erro interno }
...